docs: rephrase a bunch of things, add more links
Digging into Engine's code didn't reveal much. There truly isn't much to write
about Engine's API, it is indeed very small. Admittedly, I am not too familiar
with Babel macros, so may be I missed things because of that.

The best suggestion I have at the moment is to release the documentation to some
fresh eyes, and have them develop a small-ish application. Get the feedback and
improve the documentation based on tat.